requirements.txt

requests
ruamel.yaml
